+++ src/acceptance/org/smoothbuild/acceptance/lang/DefaultValueTest.java
@@ -1,5 +1,6 @@
 package org.smoothbuild.acceptance.lang;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.smoothbuild.acceptance.ArrayMatcher.isArrayWith;
 import static org.smoothbuild.acceptance.FileContentMatcher.hasContent;
 import static org.testory.Testory.then;
@@ -42,10 +43,27 @@ public class DefaultValueTest extends AcceptanceTestCase {
   }
 
   @Test
-  public void default_value_for_array_is_empty_array() throws Exception {
+  public void default_value_for_string_array_is_empty_array() throws Exception {
     givenScript("result: stringArrayIdentity();");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
     then(artifact("result"), isArrayWith());
+  }
+
+  @Test
+  public void default_value_for_nothing_array_is_empty_array() throws Exception {
+    givenScript("result: nothingArrayIdentity();");
+    whenSmoothBuild("result");
+    thenFinishedWithSuccess();
+    then(artifact("result"), isArrayWith());
+  }
+
+  @Test
+  public void default_value_for_nothing_doesnt_exist() throws Exception {
+    givenScript("result: nothingIdentity();");
+    whenSmoothBuild("result");
+    thenFinishedWithError();
+    then(output(), containsString("build.smooth:1: error: Parameter 'nothing' has to be "
+        + "assigned explicitly as type 'Nothing' doesn't have default value."));
   }
 }