+++ src/main/java/org/smoothbuild/lang/function/def/ArgumentExpressionCreator.java
@@ -69,8 +69,17 @@ public class ArgumentExpressionCreator {
     Map<String, Expression> argumentExpressions = convert(argumentMap);
     for (Parameter parameter : parametersPool.allOptional()) {
       Value value = parameter.type().defaultValue(valuesDb);
-      Expression expression = new ValueExpression(value, codeLocation);
-      argumentExpressions.put(parameter.name(), expression);
+      if (value == null) {
+        console.error(codeLocation, "Parameter '" + parameter.name()
+            + "' has to be assigned explicitly as type '" + parameter.type().name()
+            + "' doesn't have default value.");
+      } else {
+        Expression expression = new ValueExpression(value, codeLocation);
+        argumentExpressions.put(parameter.name(), expression);
+      }
+    }
+    if (console.isErrorReported()) {
+      return null;
     }
 
     return sortAccordingToParametersOrder(argumentExpressions, function);